backend/requirements.txt
python-dotenv >= 1.2.1
flask >= 3.1.2
flask_cors >= 6.0.2
